Working version of multi_cached decorator (#19)
multi_cached decorator allows to decorate functions that return dict-like objects in order to cache those key/values easily.

If the decorated function doesn't receive a `keys` argument, all values will be queried to the original source (meaning that the cache won't be used). If the `keys` argument is received, only the non existing keys from the cache will be queried while the existing ones will be retrieved from the cache.