Use connection that doesn't belong to the pool.
In previous versions, connections belonging to the pool were being used for
the optimistic locking. This is dangerous because redis uses the connection
to distinguish wether to apply the optimistic locking conditions or not.

In this version a new connection is acquired. This means the instance
can end up using more connections than expected at some point but they
are terminated once the operation is done (i.e. functions decorated with
@cached).