aiocache settings now are managed by Settings
before the approach was using globals and it was
messy because of adding proxies to set variables
and retrieving them. Now everything goes through
the Settings class which does all the work.

Also changed the contract of how to set defaults,
now the functions are called ``set_cache``,
``set_serializer`` and ``set_plugins`` and all
require the arg specifying the class.